QoL - Organize patches by what they are doing
